cran prep

+++ tests/testthat/setup.R
@@ -1,42 +1,55 @@
-library(Eunomia)
-connectionDetails <- getEunomiaConnectionDetails()
-
-if (grepl("testthat", getwd())) {
-  cohortPath <- "cohorts"
-} else {
-  cohortPath <- file.path("tests", "testthat", "cohorts")
+onCran <- function() {
+  notCran <- Sys.getenv("NOT_CRAN")
+  if (identical(notCran, "")) {
+    !interactive()
+  } else {
+    !isTRUE(as.logical(notCran))
+  }
 }
 
-creationFile <- file.path(cohortPath, "CohortsToCreate.csv")
-cohortDefinitionSet <- CohortGenerator::getCohortDefinitionSet(
-  settingsFileName = creationFile,
-  sqlFolder = cohortPath,
-  jsonFolder = cohortPath,
-  cohortFileNameValue = c("cohortId")
-)
+runIntegrationTests <- !onCran() &&
+  identical(tolower(Sys.getenv("PHEVALUATOR_RUN_INTEGRATION_TESTS")), "true")
 
-cohortTableNames <- CohortGenerator::getCohortTableNames(cohortTable = "cohort")
+if (runIntegrationTests) {
+  connectionDetails <- Eunomia::getEunomiaConnectionDetails()
 
-CohortGenerator::createCohortTables(
-  connectionDetails = connectionDetails,
-  cohortTableNames = cohortTableNames,
-  cohortDatabaseSchema = "main",
-  incremental = FALSE
-)
+  if (grepl("testthat", getwd())) {
+    cohortPath <- "cohorts"
+  } else {
+    cohortPath <- file.path("tests", "testthat", "cohorts")
+  }
+
+  creationFile <- file.path(cohortPath, "CohortsToCreate.csv")
+  cohortDefinitionSet <- CohortGenerator::getCohortDefinitionSet(
+    settingsFileName = creationFile,
+    sqlFolder = cohortPath,
+    jsonFolder = cohortPath,
+    cohortFileNameValue = c("cohortId")
+  )
+
+  cohortTableNames <- CohortGenerator::getCohortTableNames(cohortTable = "cohort")
+
+  CohortGenerator::createCohortTables(
+    connectionDetails = connectionDetails,
+    cohortTableNames = cohortTableNames,
+    cohortDatabaseSchema = "main",
+    incremental = FALSE
+  )
 
 
-CohortGenerator::generateCohortSet(
-  connectionDetails = connectionDetails,
-  cdmDatabaseSchema = "main",
-  cohortDatabaseSchema = "main",
-  cohortTableNames = cohortTableNames,
-  cohortDefinitionSet = cohortDefinitionSet,
-  incremental = FALSE
-)
+  CohortGenerator::generateCohortSet(
+    connectionDetails = connectionDetails,
+    cdmDatabaseSchema = "main",
+    cohortDatabaseSchema = "main",
+    cohortTableNames = cohortTableNames,
+    cohortDefinitionSet = cohortDefinitionSet,
+    incremental = FALSE
+  )
 
-withr::defer(
-  {
-    unlink(connectionDetails$server())
-  },
-  testthat::teardown_env()
-)
+  withr::defer(
+    {
+      unlink(connectionDetails$server())
+    },
+    testthat::teardown_env()
+  )
+}
